+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## v1.0.23 2023-06-11
+## v1.1.0 2023-06-13
+
+- 🆕 Added icon to the Status Bar
+- ⚒️ Refactored the Status Bar
+- 🆕 Added a new command to chnage the Bootstrap version
+- 🐛 Bug fixes
+
+## v1.0.3 2023-06-11
 
 - Bug fixes